ECO-2670: Reverted deprecations, fixed lisence in tests.
