+++ src/SprykerEco/Zed/ArvatoRss/Business/Api/Mapper/StoreOrderRequestMapper.php
@@ -7,17 +7,17 @@
 
 namespace SprykerEco\Zed\ArvatoRss\Business\Api\Mapper;
 
+use Generated\Shared\Transfer\ArvatoRssIdentificationRequestTransfer;
 use Generated\Shared\Transfer\ArvatoRssOrderTransfer;
 use Generated\Shared\Transfer\ArvatoRssStoreOrderRequestTransfer;
 use Generated\Shared\Transfer\OrderMapperTransfer;
 use Generated\Shared\Transfer\OrderTransfer;
+use SprykerEco\Shared\ArvatoRss\ArvatoRssApiConfig;
 use SprykerEco\Zed\ArvatoRss\ArvatoRssConfig;
 use SprykerEco\Zed\ArvatoRss\Business\Api\Mapper\Aspect\IdentificationMapperInterface;
 use SprykerEco\Zed\ArvatoRss\Business\Api\Mapper\Aspect\OrderMapperInterface;
+use SprykerEco\Zed\ArvatoRss\Business\Reader\ArvatoRssReaderInterface;
 
-/**
- * @deprecated Use `\SprykerEco\Zed\ArvatoRss\Business\Api\Mapper\StoreOrderCallRequestMapper` instead.
- */
 class StoreOrderRequestMapper implements StoreOrderRequestMapperInterface
 {
     /**
@@ -36,18 +36,26 @@ class StoreOrderRequestMapper implements StoreOrderRequestMapperInterface
     protected $config;
 
     /**
+     * @var \SprykerEco\Zed\ArvatoRss\Business\Reader\ArvatoRssReaderInterface|null
+     */
+    protected $reader;
+
+    /**
      * @param \SprykerEco\Zed\ArvatoRss\Business\Api\Mapper\Aspect\IdentificationMapperInterface $identificationMapper
      * @param \SprykerEco\Zed\ArvatoRss\Business\Api\Mapper\Aspect\OrderMapperInterface $orderMapper
      * @param \SprykerEco\Zed\ArvatoRss\ArvatoRssConfig $config
+     * @param \SprykerEco\Zed\ArvatoRss\Business\Reader\ArvatoRssReaderInterface|null $reader
      */
     public function __construct(
         IdentificationMapperInterface $identificationMapper,
         OrderMapperInterface $orderMapper,
-        ArvatoRssConfig $config
+        ArvatoRssConfig $config,
+        ?ArvatoRssReaderInterface $reader = null
     ) {
         $this->identificationMapper = $identificationMapper;
         $this->orderMapper = $orderMapper;
         $this->config = $config;
+        $this->reader = $reader;
     }
 
     /**
@@ -55,11 +63,13 @@ class StoreOrderRequestMapper implements StoreOrderRequestMapperInterface
      *
      * @return \Generated\Shared\Transfer\ArvatoRssStoreOrderRequestTransfer
      */
-    public function mapOrderToRequestTransfer(OrderTransfer $orderTransfer)
+    public function mapOrderToRequestTransfer(OrderTransfer $orderTransfer): ArvatoRssStoreOrderRequestTransfer
     {
         $requestTransfer = new ArvatoRssStoreOrderRequestTransfer();
 
         $identification = $this->identificationMapper->map();
+        $identification = $this->updateIdentification($orderTransfer, $identification);
+
         $order = $this->orderMapper->map(
             $this->createOrderMapperTransfer($orderTransfer)
         );
@@ -108,5 +118,30 @@ class StoreOrderRequestMapper implements StoreOrderRequestMapperInterface
         $transfer->setCustomerIsGuest($customer ? $customer->getIsGuest() : true);
 
         return $transfer;
+    }
+
+    /**
+     * @param \Generated\Shared\Transfer\OrderTransfer $orderTransfer
+     * @param \Generated\Shared\Transfer\ArvatoRssIdentificationRequestTransfer $identification
+     *
+     * @return \Generated\Shared\Transfer\ArvatoRssIdentificationRequestTransfer
+     */
+    protected function updateIdentification(
+        OrderTransfer $orderTransfer,
+        ArvatoRssIdentificationRequestTransfer $identification
+    ): ArvatoRssIdentificationRequestTransfer {
+        if ($this->reader === null) {
+            return $identification;
+        }
+
+        $arvatoRssApiCallLogTransfer = $this->reader
+            ->getApiLogByOrderReferenceAndType(
+                $orderTransfer->getOrderReference(),
+                ArvatoRssApiConfig::TRANSACTION_TYPE_RISK_CHECK
+            );
+
+        $identification->setCommunicationToken($arvatoRssApiCallLogTransfer->getCommunicationToken());
+
+        return $identification;
     }
 }
