+++ tests/SprykerEcoTest/Zed/ArvatoRss/_support/Mock/Api/Adapter/AdapterFactoryMock.php
@@ -2,7 +2,7 @@
 
 /**
  * MIT License
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
+ * For full license information, please view the LICENSE file that was distributed with this source code.
  */
 
 namespace SprykerEcoTest\Zed\ArvatoRss\Mock\Api\Adapter;
@@ -20,7 +20,7 @@ class AdapterFactoryMock extends AdapterFactory
     {
         return new StoreOrderCallMock(
             $this->createRequestHeaderConverter(),
-            $this->createArvatoRssApiCallLogger()
+            $this->createApiCallLogger()
         );
     }
 }
