+++ tests/SprykerEcoTest/Zed/ArvatoRss/_support/Mock/Api/Adapter/SoapApiAdapterMock.php
@@ -2,7 +2,7 @@
 
 /**
  * MIT License
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
+ * For full license information, please view the LICENSE file that was distributed with this source code.
  */
 
 namespace SprykerEcoTest\Zed\ArvatoRss\Mock\Api\Adapter;
@@ -13,7 +13,7 @@ use SprykerEco\Zed\ArvatoRss\Business\Api\Adapter\SoapApiAdapter;
 
 class SoapApiAdapterMock extends SoapApiAdapter
 {
-    const WSDL_PATH = __DIR__ . "/../../../../../../../../src/SprykerEco/Zed/ArvatoRss/Business/Api/Etc/risk-solution-services.v2.1.wsdl";
+    public const WSDL_PATH = __DIR__ . "/../../../../../../../../src/SprykerEco/Zed/ArvatoRss/Business/Api/Etc/risk-solution-services.v2.1.wsdl";
 
     /**
      * @param \Generated\Shared\Transfer\ArvatoRssRiskCheckRequestTransfer $requestTransfer
