+++ tests/SprykerEcoTest/Zed/ArvatoRss/_support/Mock/MoneyFacadeMock.php
@@ -2,7 +2,7 @@
 
 /**
  * MIT License
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
+ * For full license information, please view the LICENSE file that was distributed with this source code.
  */
 
 namespace SprykerEcoTest\Zed\ArvatoRss\Mock;
@@ -11,8 +11,8 @@ use SprykerEco\Zed\ArvatoRss\Dependency\Facade\ArvatoRssToMoneyInterface;
 
 class MoneyFacadeMock implements ArvatoRssToMoneyInterface
 {
-    const VALUE_DECIMAL = 100.00;
-    const VALUE_INT = 10000;
+    public const VALUE_DECIMAL = 100.00;
+    public const VALUE_INT = 10000;
 
     /**
      * @param int $value
